+++ frontend/src/metabase/databases/components/DatabaseHelpCard/DatabaseHelpCard.tsx
@@ -1,18 +1,20 @@
 import { jt, t } from "ttag";
-import Settings from "metabase/lib/settings";
+import { useSelector } from "metabase/lib/redux";
+import { getDocsUrl, getSetting } from "metabase/selectors/settings";
 import HelpCard from "metabase/components/HelpCard";
 import ExternalLink from "metabase/core/components/ExternalLink";
 
 export interface DatabaseHelpCardProps {
   className?: string;
-  isHosted?: boolean;
 }
 
-const DatabaseHelpCard = ({
+export const DatabaseHelpCard = ({
   className,
-  isHosted,
 }: DatabaseHelpCardProps): JSX.Element => {
-  const docsUrl = Settings.docsUrl("databases/connecting");
+  const docsUrl = useSelector(state =>
+    getDocsUrl(state, { page: "databases/connecting" }),
+  );
+  const isHosted = useSelector(state => getSetting(state, "is-hosted?"));
 
   return (
     <HelpCard
@@ -34,6 +36,3 @@ const DatabaseHelpCard = ({
     </HelpCard>
   );
 };
-
-// eslint-disable-next-line import/no-default-export -- deprecated usage
-export default DatabaseHelpCard;
