+++ frontend/src/metabase/databases/components/DatabaseHelpCard/index.ts
@@ -1,2 +1,1 @@
-// eslint-disable-next-line import/no-default-export -- deprecated usage
-export { default } from "./DatabaseHelpCard";
+export * from "./DatabaseHelpCard";
