--- frontend/src/metabase/databases/containers/DatabaseHelpCard/index.ts
@@ -1,2 +1,0 @@
-// eslint-disable-next-line import/no-default-export -- deprecated usage
-export { default } from "./DatabaseHelpCard";
