+++ frontend/src/metabase/setup/components/DatabaseHelp/DatabaseHelp.tsx
@@ -1,5 +1,5 @@
 import { useSelector } from "metabase/lib/redux";
-import DatabaseHelpCard from "metabase/databases/containers/DatabaseHelpCard";
+import { DatabaseHelpCard } from "metabase/databases/components/DatabaseHelpCard";
 import { DATABASE_STEP } from "../../constants";
 import { getDatabaseEngine, getIsStepActive } from "../../selectors";
 import { SetupCardContainer } from "../SetupCardContainer";
